+++ pkg/providers/gcp/provider_sim_test.go
@@ -41,8 +41,8 @@ switches:
 blocks:
 - switch: tor
   nodes: ["11"]
-  labels:
-    network.topology.nvidia.com/accelerator: nvl1
+  annotations:
+    accelerator.topology.test/domain: nvl1
 `
 
 	clusterModel = `
@@ -56,12 +56,12 @@ switches:
 blocks:
 - switch: tor1
   nodes: ["11","12"]
-  labels:
-    network.topology.nvidia.com/accelerator: nvl1
+  annotations:
+    accelerator.topology.test/domain: nvl1
 - switch: tor2
   nodes: ["21","22"]
-  labels:
-    network.topology.nvidia.com/accelerator: nvl2
+  annotations:
+    accelerator.topology.test/domain: nvl2
 `
 )
 
@@ -124,8 +124,8 @@ switches:
 blocks:
 - switch: tor
   nodes: [n11]
-  labels:
-    network.topology.nvidia.com/accelerator: nvl1
+  annotations:
+    accelerator.topology.test/domain: nvl1
 `,
 			instances: []topology.ComputeInstances{
 				{